Lucas

+++ Game/src/Fase/Deserto.cpp
@@ -14,7 +14,7 @@ namespace Fases {
 		background.setTexture(&background_texture);
 		background_auxiliar.setSize(sf::Vector2f(ger_grafico->getWindowSize()));
 		background_auxiliar.setTexture(&background_texture);
-		background_auxiliar.setPosition(ger_grafico->getWindowSize().x, 0.f);
+		background_auxiliar.setPosition((float) ger_grafico->getWindowSize().x, 0.f);
 	}
 
 	Deserto::~Deserto()
